+++ plugin-packs/postcss-preset-env/src/lib/ids-by-execution-order.mjs
@@ -1,6 +1,5 @@
 // ids ordered by required execution, then alphabetically
 export default [
-	'cascade-layers',
 	'custom-media-queries',
 	'custom-properties',
 	'environment-variables', // run environment-variables here to access transpiled custom media params and properties
@@ -37,4 +36,5 @@ export default [
 	'system-ui-font-family',
 	'stepped-value-functions',
 	'trigonometric-functions',
+	'cascade-layers',
 ];